# Improve space usage and time of decoding.