--- a/proto-lens/src/Data/ProtoLens/Encoding.hs
+++ b/proto-lens/src/Data/ProtoLens/Encoding.hs
@@ -8,7 +8,9 @@
 --
 -- TODO: Currently all operations are on strict ByteStrings;
 -- we should try to generalize to lazy Bytestrings as well.
+{-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE GADTs #-}
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE PatternGuards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 module Data.ProtoLens.Encoding(
@@ -23,7 +25,7 @@ import Data.ProtoLens.Encoding.Bytes
 import Data.ProtoLens.Encoding.Wire
 
 import Control.Applicative ((<|>), (<$>))
-import Control.Monad (foldM)
+import Control.Monad (foldM, join)
 import Data.Attoparsec.ByteString as Parse
 import Data.Bool (bool)
 import Data.Text.Encoding (encodeUtf8, decodeUtf8')
@@ -35,6 +37,7 @@ import qualified Data.ByteString.Lazy as L
 import Data.Monoid (mconcat, mempty)
 import Data.Foldable (foldMap, toList, foldl')
 import Lens.Family2 (set, over, (^.), (&))
+import Data.Functor.Identity (Identity(..))
 
 -- TODO: We could be more incremental when parsing/encoding length-based fields,
 -- rather than forcing the whole thing.  E.g., for encoding we're doing extra
@@ -43,9 +46,41 @@ import Lens.Family2 (set, over, (^.), (&))
 -- | Decode a message from its wire format.  Returns 'Left' if the decoding
 -- fails.
 decodeMessage :: Message msg => B.ByteString -> Either String msg
-decodeMessage input =
-    parseOnly (Parse.manyTill getTaggedValue endOfInput) input
-        >>= taggedValuesToMessage
+decodeMessage input = parseOnly (parseMessage end) input
+  where
+    end = endOfInput *> pure Nothing
+
+-- | Parse a message with the given ending delimiter (which will be EndGroup in
+-- the case of a group, and end-of-input otherwise).  The ending delimiter
+-- Parser returns (Just err) if it encountered an unrecoverable error; in
+-- particular, if it found an EndGroup with an unexpected tag number.
+parseMessage :: forall msg . Message msg => Parser (Maybe String) -> Parser msg
+parseMessage end = do
+    (msg, unsetFields) <- join $ loop def requiredFields
+    if Map.null unsetFields
+        then return $ reverseRepeatedFields fields msg
+        else fail $ "Missing required fields "
+                        ++ show (map fieldDescriptorName
+                                    $ Map.elems $ unsetFields)
+  where
+    fields = fieldsByTag descriptor
+    requiredFields = Map.filter isRequired fields
+    loop :: msg -> Map.Map Tag (FieldDescriptor msg)
+            -> Parser (Parser (msg, Map.Map Tag (FieldDescriptor msg)))
+    -- Note that the following won't work:
+    -- end >>= \case {Just err -> fail err; ...}
+    -- since otherwise the below <|> will cause us to ignore an error that's
+    -- supposed to be unrecoverable.
+    loop msg unsetFields = (end >>= pure . \case
+                                Nothing -> pure (msg, unsetFields)
+                                Just err -> fail err)
+                <|> do
+                    tv@(TaggedValue tag _) <- getTaggedValue
+                    case Map.lookup (Tag tag) fields of
+                        Nothing -> loop msg unsetFields
+                        Just field -> do
+                            !msg' <- parseAndAddField msg field tv
+                            loop msg' $! Map.delete (Tag tag) unsetFields
 
 -- | Decode a message from its wire format.  Throws an error if the decoding
 -- fails.
@@ -54,30 +89,6 @@ decodeMessageOrDie bs = case decodeMessage bs of
     Left e -> error $ "decodeMessageOrDie: " ++ e
     Right x -> x
 
--- | Convert a sequence of parsed key-value pairs into a Message via its
--- descriptor. Will fail if any of the key-value pairs do not match those
--- expected by the field descriptors.
-taggedValuesToMessage :: Message msg => [TaggedValue] -> Either String msg
-taggedValuesToMessage tvs
-    | missing <- missingFields fields tvs, not $ null missing
-        = Left $ "Missing required fields " ++ show missing
-    | otherwise = reverseRepeatedFields fields <$> result
-  where
-    addTaggedValue msg tv@(TaggedValue tag _) =
-        case Map.lookup (Tag tag) fields of
-            Nothing -> return msg
-            Just field -> parseAndAddField msg field tv
-    fields = fieldsByTag descriptor
-    result = foldM addTaggedValue def tvs
-
-missingFields :: Map.Map Tag (FieldDescriptor msg) -> [TaggedValue] -> [String]
-missingFields fields
-    = map fieldDescriptorName
-        . Map.elems
-        . foldl' (\m (TaggedValue t _) -> Map.delete (Tag t) m) requiredFields
-  where
-    requiredFields = Map.filter isRequired fields
-
 runEither :: Either String a -> Parser a
 runEither (Left x) = fail x
 runEither (Right x) = return x
@@ -85,38 +96,47 @@ runEither (Right x) = return x
 parseAndAddField :: msg
                  -> FieldDescriptor msg
                  -> TaggedValue
-                 -> Either String msg
+                 -> Parser msg
 parseAndAddField
-    msg
+    !msg
     (FieldDescriptor name typeDescriptor accessor)
-    (TaggedValue tag (WireValue wt val))
-    = case fieldWireType typeDescriptor of
-        FieldWireType fieldWt _ get -> let
-          getSimpleVal = do
-              Equal <- equalWireTypes name fieldWt wt
-              get val
+    (TaggedValue tag (WireValue wt val)) = let
+          getSimpleVal = case fieldWireType typeDescriptor of
+                            GroupFieldType -> do
+                                Equal <- equalWireTypes name StartGroup wt
+                                parseMessage (endOfGroup name tag)
+                            FieldWireType fieldWt _ get -> do
+                                Equal <- equalWireTypes name fieldWt wt
+                                runEither $ get val
           -- Get a block of packed values, reversed.
-          getPackedVals = do
+          getPackedVals = case fieldWireType typeDescriptor of
+            GroupFieldType -> fail "Groups can't be packed"
+            FieldWireType fieldWt _ get -> do
               Equal <- equalWireTypes name Lengthy wt
-              let getElt = getWireValue fieldWt tag >>= runEither . get
-              parseOnly (manyReversedTill getElt endOfInput) val
+              let getElt = do
+                        wv <- getWireValue fieldWt tag
+                        x <- runEither $ get wv
+                        return $! x
+              runEither $ parseOnly (manyReversedTill getElt endOfInput) val
           in case accessor of
               PlainField _ f -> do
-                  x <- getSimpleVal
-                  return $ set f x msg
+                  !x <- getSimpleVal
+                  return $! set f x msg
               OptionalField f -> do
-                  x <- getSimpleVal
-                  return $ set f (Just x) msg
+                  !x <- getSimpleVal
+                  return $! set f (Just x) msg
               RepeatedField Unpacked f -> do
-                  x <- getSimpleVal
-                  return $ over f (x:) msg
+                  !x <- getSimpleVal
+                  return $! over f (\(!xs) -> x:xs) msg
               RepeatedField Packed f -> do
                   xs <- getPackedVals
-                  return $ over f (xs++) msg
+                  return $! over f (\(!ys) -> xs++ys) msg
               MapField keyLens valueLens f -> do
                   entry <- getSimpleVal
-                  return $ over f
-                      (Map.insert (entry ^. keyLens) (entry ^. valueLens))
+                  let !key = entry ^. keyLens
+                  let !value = entry ^. valueLens
+                  return $! over f
+                      (Map.insert key value)
                       msg
 
 -- | Run the parser zero or more times, until the "end" parser succeeds.
@@ -137,42 +157,51 @@ buildMessage msg = foldMap putTaggedValue (messageToTaggedValues msg)
 -- | Encode a message as a sequence of key-value pairs.
 messageToTaggedValues :: Message msg => msg -> [TaggedValue]
 messageToTaggedValues msg = mconcat
-    [ map (TaggedValue t) (messageFieldToVals fieldDescr msg)
+    [ messageFieldToVals t fieldDescr msg
     | (Tag t, fieldDescr) <- Map.toList (fieldsByTag descriptor)
     ]
 
-messageFieldToVals :: FieldDescriptor msg -> msg -> [WireValue]
-messageFieldToVals (FieldDescriptor _ typeDescriptor accessor) msg =
-    case fieldWireType typeDescriptor of
-        FieldWireType wt convert _ -> case accessor of
+messageFieldToVals :: Int -> FieldDescriptor a -> a -> [TaggedValue]
+messageFieldToVals tag (FieldDescriptor _ typeDescriptor accessor) msg =
+    let
+        embed src
+            = case fieldWireType typeDescriptor of
+                FieldWireType wt convert _
+                    -> [TaggedValue tag $ WireValue wt
+                                            (convert src)]
+                GroupFieldType
+                    -> TaggedValue tag (WireValue StartGroup ())
+                            : messageToTaggedValues src
+                                ++ [TaggedValue tag $ WireValue EndGroup ()]
+        embedPacked src
+            = case fieldWireType typeDescriptor of
+                GroupFieldType -> fail "packged GroupFieldType"
+                FieldWireType wt convert _ -> let
+                    v = L.toStrict $ toLazyByteString
+                        $ mconcat [putWireValue wt (convert x) | x <- src]
+                    in [TaggedValue tag $ WireValue Lengthy v]
+    in case accessor of
             PlainField d f
                 | Optional <- d, src == fieldDefault -> []
-                | otherwise -> [WireValue wt (convert src)]
+                | otherwise -> embed src
               where src = msg ^. f
             OptionalField f -> case msg ^. f of
-                Just src -> [WireValue wt (convert src)]
-                _ -> mempty
+                                Just src -> embed src
+                                _ -> mempty
             RepeatedField Unpacked f
-                -> [ WireValue wt (convert src)
-                   | src <- toList (msg ^. f)
-                   ]
+                -> mconcat [embed src | src <- toList (msg ^. f)]
             RepeatedField Packed f
-                -> [WireValue Lengthy v]
-                     where v = L.toStrict $ toLazyByteString
-                               $ mconcat
-                                 [ putWireValue wt (convert src)
-                                 | src <- toList (msg ^. f)
-                                 ]
+                -> embedPacked (msg ^. f)
             MapField keyLens valueLens f ->
-                [ WireValue wt v
+                mconcat [ embed entry
                 | (key, value) <- Map.toList (msg ^. f)
                 , let entry = def & set keyLens key & set valueLens value
-                , let v = convert entry
                 ]
 
 data FieldWireType value where
     FieldWireType :: WireType w -> (value -> w) -> (w -> Either String value)
                   -> FieldWireType value
+    GroupFieldType :: Message value => FieldWireType value
 
 fieldWireType :: FieldTypeDescriptor value -> FieldWireType value
 -- TODO: Don't let toEnum crash on unknown enum values.
@@ -199,11 +228,22 @@ fieldWireType FloatField = simpleFieldWireType Fixed32 floatToWord wordToFloat
 fieldWireType DoubleField = simpleFieldWireType Fixed64
                                 doubleToWord wordToDouble
 fieldWireType StringField = FieldWireType Lengthy encodeUtf8
-                                                  (stringizeError . decodeUtf8')
+                                    (stringizeError . decodeUtf8')
 fieldWireType BytesField = identityFieldWireType Lengthy
-fieldWireType MessageField = FieldWireType Lengthy encodeMessage decodeMessage
-fieldWireType GroupField =
-    FieldWireType StartGroup messageToTaggedValues taggedValuesToMessage
+fieldWireType MessageField = FieldWireType Lengthy encodeMessage
+                                decodeMessage
+fieldWireType GroupField = GroupFieldType
+
+-- TODO: check it's the same tag, and if not then fail the *whole* computation.
+endOfGroup :: String -> Int -> Parser (Maybe String)
+endOfGroup name tag = do
+    TaggedValue tag' (WireValue wt _) <- getTaggedValue
+    Equal <- equalWireTypes name EndGroup wt
+    return $ if tag == tag'
+                then Nothing
+                else Just $ "endOfGroup: unexpected tag " ++ show tag'
+                                            ++ " for field " ++ show name
+                                                ++ ", expected " ++ show tag
 
 -- | Helper function to define a field type whose decoding operation can't fail.
 simpleFieldWireType :: WireType w -> (value -> w) -> (w -> value)