Motivated by #58, it turned out there were a few space leaks in the decoder:
- Made all message fields strict
- Add bang patterns to the decoding code
- Don't generate an intermediate list of TaggedValues.

These changes had the happy side effect of significantly speeding up our
decoding benchmarks.  Additionally the "decode/whnf" and "decode/nf" tests now
take the same amount of time, indicating that we do fully force the messages.

|                                              |              before  |                after |
| -------------------------------------------- | -------------------- | -------------------- |
|nested:flat(68kB)/encode                      |  3.547ms (+/- 0.009) |  2.559ms (+/- 0.003) |
|nested:flat(68kB)/decode/whnf                 | 16.304ms (+/- 0.295) |  7.225ms (+/- 0.039) |
|nested:flat(68kB)/decode/nf                   | 19.417ms (+/- 0.198) |  7.283ms (+/- 0.044) |
|nested:nested(87kB)/encode                    | 10.670ms (+/- 0.279) |  7.791ms (+/- 0.040) |
|nested:nested(87kB)/decode/whnf               | 18.310ms (+/- 0.347) | 13.749ms (+/- 0.023) |
|nested:nested(87kB)/decode/nf                 | 21.977ms (+/- 0.185) | 13.780ms (+/- 0.043) |
|int-packing:int32-unpacked(19kB)/encode       |  1.319ms (+/- 0.003) |  0.704ms (+/- 0.011) |
|int-packing:int32-unpacked(19kB)/decode/whnf  |  8.300ms (+/- 0.337) |  3.246ms (+/- 0.005) |
|int-packing:int32-unpacked(19kB)/decode/nf    |  9.772ms (+/- 0.302) |  3.269ms (+/- 0.006) |
|int-packing:int32-packed(10003B)/encode       |  0.869ms (+/- 0.003) |  0.775ms (+/- 0.002) |
|int-packing:int32-packed(10003B)/decode/whnf  |  1.809ms (+/- 0.048) |  1.670ms (+/- 0.002) |
|int-packing:int32-packed(10003B)/decode/nf    |  3.372ms (+/- 0.129) |  1.707ms (+/- 0.002) |
|unused-fields:no-unused(19kB)/encode          |  1.341ms (+/- 0.004) |  0.693ms (+/- 0.004) |
|unused-fields:no-unused(19kB)/decode/whnf     |  8.253ms (+/- 0.076) |  3.125ms (+/- 0.015) |
|unused-fields:no-unused(19kB)/decode/nf       |  9.713ms (+/- 0.072) |  3.148ms (+/- 0.016) |
|unused-fields:with-unused(19kB)/encode        |  1.363ms (+/- 0.004) |  0.720ms (+/- 0.035) |
|unused-fields:with-unused(19kB)/decode/whnf   | 12.224ms (+/- 0.096) |  3.599ms (+/- 0.099) |
|unused-fields:with-unused(19kB)/decode/nf     | 13.383ms (+/- 0.233) |  3.608ms (+/- 0.007) |